% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_stats.R
\name{bootstrap_maia}
\alias{bootstrap_maia}
\title{bootstrapping maia norm data}
\usage{
bootstrap_maia(n = 50, remove_diff = TRUE)
}
\arguments{
\item{n}{number of bootstrap samples}

\item{remove_diff}{logical. TRUE will remove cr_diff data (default)}
}
\value{
List
}
\description{
creates n bootstrap samples of the original norm_data
}
\details{
Returns list of linear models
}
\seealso{
Other stat functions: 
\code{\link{mpstats_manual}()},
\code{\link{mpstats_single}()},
\code{\link{mpstats}()}
}
\author{
tjebo
}
\concept{stat functions}
