% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_stats.R
\name{calc_bebie}
\alias{calc_bebie}
\title{calculates bebie statistics}
\usage{
calc_bebie(testdata)
}
\arguments{
\item{testdata}{test for which field variation will be calculated}
}
\value{
List
}
\description{
stats for bebie curve
}
\author{
tjebo
}
\concept{prediction functions}
