% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{denniss_nest}
\alias{denniss_nest}
\title{Norm data 2}
\format{
A nested tibble with:
\describe{
\item{patID}{Anonymized patient identifier}
\item{sex}{Sex of patient (m = male, f = female)}
\item{eye}{Left or right eye of patient (r = right, l = left)}
}
}
\source{
https://www.sciencedirect.com/science/article/pii/S2352340916304978
accompagnying paper: http://iovs.arvojournals.org/article.aspx?articleid=2571342
}
\description{
Norm data obtained from healthy controls in a UK center
by Jonathan Denniss.\preformatted{**TO BE CONTINUED**
}
}
