% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_maia.R
\name{show_tgz}
\alias{show_tgz}
\alias{unpack_tgz}
\alias{create_ID_dirs}
\alias{get_json}
\title{data extraction}
\usage{
show_tgz(folder = getwd())

unpack_tgz(tgz_dirs)

create_ID_dirs(tgz_dirs)

get_json(ID_dirs)
}
\arguments{
\item{folder}{source folder. Default: workdirectory}

\item{tgz_dirs}{character vector of tgz folders}
}
\description{
For data extraction of MAIA files

Extracts name, dob and and sex from MAIA observers.
This data is stored in json files attached to the microperimetry files.
It is generally not recommended to use this function explicitly in order to
keep personal identifiable data as safe as possible.
}
\keyword{internal}
