% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_stats.R
\name{mpstats}
\alias{mpstats}
\title{Calculates microperimetry statistics.}
\usage{
mpstats(testdata, digits = 2)
}
\arguments{
\item{testdata}{microperimetry test data}

\item{digits}{Digits to be rounded to}
}
\value{
Matrix
}
\description{
Calculates microperimetry statistics from a test or list of tests.
Includes mean sensitivity (mean_sens), mean deviation (mean_dev),
pattern standard deviation (psd).
You can also pass a compare object directly (from compare function),
thus saving time for duplicate kringing.
}
\details{
\strong{mean_dev} and \strong{psd} are estimated.
This is, because there is no 'real mean' and, more importantly,
no "real variance" of normal values for each given location.
Mean and variance are estimated based on kringing (spatial interpolation)
of \emph{predicted (!)} values. The predicted values are given by simple linear
regression for each location, based on the norm data
It includes age, sex and lens status as covariates.
The variance used for estimation is equal to the \emph{prediction intervals}
from those models, assuming homoscedasticity in a non-weighted model
}
\seealso{
Other stat functions: 
\code{\link{bootstrap_maia}()},
\code{\link{mpstats_manual}()},
\code{\link{mpstats_single}()}
}
\author{
tjebo
}
\concept{stat functions}
