% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perimetry_package.R
\docType{package}
\name{perimetry}
\alias{perimetry}
\title{perimetry}
\description{
An R framework for perimetry analysis. Currently only
supporting MAIA device, but the architecture of the package allows
for extension to other devices.
}
\section{Extracting data}{

Extracting data from MAIA raw (tgz) or .txt in a single call.
}

\section{Statistics}{

Spatially-ignorant metrics (will be spatially-weighted)
\itemize{
\item Mean Sensitivity (dB)
\item Mean Deviation (dB)
\item PSD (dB)
\item Mean Loss (dB)
\item Sqrt. of Loss Variance (dB)
}

Spatially-adjusted metrics
\itemize{
\item Vol. (dB*sr)
\item Loss (dB*sr)
\item PSD (dB*sr)
}
}

\section{Visualisation}{

perimetry also includes functions for some more or less common visualization
of visual field data
}

\author{
Tjebo Heeren and Maximilian Pfau and ... TBC!
}
