% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pfau_nest}
\alias{pfau_nest}
\title{Norm data 1}
\format{
A nested tibble with:
\describe{
\item{patID}{Anonymized patient identifier}
\item{sex}{Sex of patient (m = male, f = female)}
\item{eye}{Left or right eye of patient (r = right, l = left)}
}
}
\usage{
data(norm_data)
}
\description{
Norm data obtained from healthy controls. Data was collected in
the University Eye Hospital Bonn by Simone Tzaridis and Maximilian Pfau.\preformatted{**TO BE CONTINUED**
}
}
