% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyfield.R
\name{tidyfield}
\alias{tidyfield}
\title{The tidyfield object}
\description{
The tidyfield object is the central object of the perimetry
package. It is designed to allow for efficient and fast usage of
perimetry data in tidy models. It is essentially a nested tibble.
}
